♻️ rename viewCollection to trackViews
