+++ packages/rum/src/domain/lifeCycle.ts
@@ -4,7 +4,7 @@ import { RawRumEvent, RumEvent } from '../types'
 import { RawRumEventV2, RumEventV2 } from '../typesV2'
 import { RequestCompleteEvent, RequestStartEvent } from './requestCollection'
 import { AutoActionCreatedEvent, AutoUserAction, CustomUserAction } from './rumEventsCollection/userActionCollection'
-import { View, ViewCreatedEvent } from './rumEventsCollection/view/viewCollection'
+import { View, ViewCreatedEvent } from './rumEventsCollection/view/trackViews'
 
 export enum LifeCycleEventType {
   ERROR_COLLECTED,
