+++ packages/rum/src/types.ts
@@ -1,7 +1,7 @@
 import { Context, ErrorContext, HttpContext, ResourceType } from '@datadog/browser-core'
 import { PerformanceResourceDetails } from './domain/rumEventsCollection/resource/resourceUtils'
 import { ActionType, UserActionMeasures } from './domain/rumEventsCollection/userActionCollection'
-import { ViewLoadingType, ViewMeasures } from './domain/rumEventsCollection/view/viewCollection'
+import { ViewLoadingType, ViewMeasures } from './domain/rumEventsCollection/view/trackViews'
 
 export enum RumEventCategory {
   USER_ACTION = 'user_action',
